+++ components/Love.js
@@ -4,7 +4,7 @@ import { IconButton, withTheme } from 'react-native-paper';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 
-import { addToFavorite, removeFromFavorite } from '../actions';
+import { addToFavorite, removeFromFavorite } from '../actions/playerState';
 
 class Love extends Component {
     constructor(props) {
@@ -67,7 +67,7 @@ class Love extends Component {
 }
 
 const mapStateToProps = state => ({
-    favorite: state.media.favorite
+    favorite: state.playerState.favorite
 });
 
 export default connect(mapStateToProps, { addToFavorite, removeFromFavorite })(withTheme(Love));
