+++ pages/offline/Album.js
@@ -2,10 +2,11 @@ import { FlatList } from 'react-native-gesture-handler';
 import * as React from 'react';
 import { withTheme, Divider, Title, List, Colors } from 'react-native-paper';
 import { connect } from 'react-redux';
-import { View, StyleSheet } from 'react-native';
+import { View, StyleSheet, RefreshControl } from 'react-native';
 import FastImage from 'react-native-fast-image'; 
 import _ from 'lodash';
-import RNAndroidAudioStore from "react-native-get-music-files";
+
+import { getOfflineAlbums } from '../../actions/mediaStore';
 
 class Album extends React.Component {
     static navigationOptions = {
@@ -15,22 +16,32 @@ class Album extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            albums: []        }
+            albums: [],
+            refreshing: false    
+        }
     }
 
-    // componentWillReceiveProps(nextProps) {
-    //     if (!_.isEmpty(nextProps.files)) {
-    //         this.setState({ files: nextProps.files });
-    //     }
-    // }
+    
+    static getDerivedStateFromProps(props, state) {
+        if (!_.isEqual(props.albums, state.albums)) {
+            return {
+                albums: props.albums,
+                refreshing: false
+            }
+        }
+        return null
+    }
+
+    fetchData = () => {
+        this.setState({
+            refreshing: true
+        })
+        this.props.getOfflineAlbums();
+    }
 
     componentDidMount() {
-        // this.props.getOfflineMedia()
-        RNAndroidAudioStore.getAlbums({})
-            .then(media => {
-                this.setState({ albums: media });
-            })
-            .catch(er => console.log(er));
+        this.props.getOfflineAlbums()
+       
     }
 
 
@@ -43,7 +54,6 @@ class Album extends React.Component {
 
         const { navigate } = this.props.navigation;
 
-        // const albums = _.uniqBy(this.state.files, 'album');
 
         if (!_.isEmpty(this.state.albums)) {
             return (
@@ -51,7 +61,12 @@ class Album extends React.Component {
                     <FlatList
                         data={this.state.albums}
                         ItemSeparatorComponent={() => <Divider inset={true} />}
-                        // onRefresh={() => this.props.getOfflineMedia()}
+                        refreshControl={
+                            <RefreshControl
+                                refreshing={this.state.refreshing}
+                                onRefresh={() => this.fetchData()}
+                            />
+                        }
                         keyExtractor={(item, index) => index.toString()}
                         renderItem={({ item }) =>
                             <List.Item
@@ -86,13 +101,14 @@ class Album extends React.Component {
     }
 }
 
-// const mapStateToProps = state => ({
-//     files: state.media.files
-// });
 
 
-// export default connect(mapStateToProps)(withTheme(Album));
-export default withTheme(Album);
+const mapStateToProps = state => ({
+    albums: state.mediaStore.albums
+});
+
+
+export default connect(mapStateToProps, { getOfflineAlbums })(withTheme(Album));
 
 const styles = StyleSheet.create({
     icons: {
